Fixed warning
git-svn-id: svn://falconpl.org/falcon/core/trunk@647 78e5f83a-d9f7-4e86-b854-dad15b5eb37a

+++ engine/core_module/inspect.cpp
@@ -164,7 +164,7 @@ void inspect_internal( VMachine *vm, const Item *elem, int32 level, int32 maxLev
                stream->writeString( temp );
                temp = "";
             }
-            if ( count == maxSize )
+            if ( count == (uint32) maxSize )
                stream->writeString( " ... " );
             stream->writeString( "]" );
          }
